Change rails new command instruction
At the Rails Girls Berlin beginners' workshop yesterday we had
problems with multiple students who had followed the installation
instructions either at home or at the installation party the
evening before.

We shouldn't instruct beginners to create a new Rails app named
railsgirls to test their installation because that app name is
also used in the app guide. It can result in people having two
apps somewhere on their machine with the same name, which causes
confusion.

This changes the name of the app created in the installation
guide. We could also consider not creating an application at all
and just running rails -v.