Split multi-line blocks in installation guide
This splits code-blocks containing multiple statements into separate
blocks, in order to make it more obvious that these commands a) should
be inputted separately and b) require the preceeding one to be finished
before the next one can be run.

Several people following the guide ran into problems with these
"multi-line" steps. Not knowing the commands and not being too familiar
with the command line, it may not be obvious on the first glance that
each line represents a separate command and that one has to wait for one
to finish before the next one can run.

Splitting this up of course looks way more verbose, but it could also
increase the readability of the guide.