Update Windows install guide
The RailsInstaller project has not been updated in a while. The last
version available installs Ruby 2.3, which has been End Of Life since
2019-03-31 (source: https://www.ruby-lang.org/en/downloads/branches/).

The last RailsGirls Amsterdam workshop we experienced problems with this
Ruby version in combination with newer gems that no longer support this
version.

For this reason, update the Windows install guide. Instead of
RailsInstaller, use the RubyInstaller, which supports newer and
supported Ruby versions. Because this installer only installs Ruby
and not all necessary components for a Rails install installation, add
extra steps to install Git and manually install SQLite.

Recommend 32-bit everywhere because Windows supports both and to explain
the difference in this guide is a bit out of scope. This way it works
for both 32 and 64-bit machines.

Since SQLite doesn't have a Windows installer - something that
automatically copies the program's files in a suitable location and adds
the location to the PATH environment variable - add the most basic
install steps for it to work. Create a new directory on the `C:` disk,
add it to the path by using a command line operation.

I chose the command-line approach to updating the system PATH so there
is just one operation to perform. Otherwise a multi-step process going
through the Windows UI is needed, along with screenshots for every step.
Something which also doesn't work well with for Windows installations in
another language.

Update the Atom section to recommend the installer, rather than having
to move the package contents to a system location again.

Also move the "Possible errors" section down to the bottom of the page
and reference it, rather than it taking up a lot of space in the middle
of the guide.