guides: update install guide for Ruby 3 (Win/Mac)
As Ruby 2.7 is end-of-life 2022-03-31 (https://endoflife.date/ruby), we
need to update the guide for Ruby 3. This commit does this for Windows
and Mac, and handles the default shell change in newer versions of
MacOS.

`webrick` is also added as a dependency, as in Ruby 3 this is no longer
bundled with Ruby itself.

ChromeOS, Linux and cloud setup are not changed - this will be a follow
up PR.