Suggest Visual Studio Code as an Atom replacement (#494)
The Atom editor has been sunset and clicking the link to the homepage
ends up on the GitHub blog post about sunsetting the editor. This is
confusing to readers and we should suggest another editor instead. VS
Code is suggested in other parts of the guide already, let's stick with
that.