Add a tools guide
Add a new tools guide based on parts of guides that were scattered
around other guides. Now these instructions are consistent.

Also make it the very first guide. The install guide started with asking
people to run commands in the Terminal, without having had an
introduction to what the Terminal app is.

I numbered it guide 0, so I didn't have to renumber all the guides right
now. That can be done later.

I added a new `_pages` directory so that this file doesn't need a date
attached to it. Let's move the other guides over at some point. I'll be
updating more guides in the coming weeks to do so.