Move OS switcher to top of commands
When the Operating System switcher is first introduced it's at the
bottom of the commands. That may lead readers to run the commands for
the wrong Operating System.

Move the Operating System switcher to the top of the Operating System
specific instructions.

Rename the "Other" label to "Mac/Linux" so it's more clear what the
other Operating Systems than Windows are.