Make code blocks full width
Make code blocks take up the full width of the page on the right hand
side. Especially on small screens like mobile device the width was very
limited resulting in a small window to view the code.

Remove the spacing on the right, this gives more visible space for text
in the code blocs.