Update the app guide instructions
I updated the instructions to apply to Rails 7, the version of Rails
most people will be installing. The instructions are also a little more
incremental, for the first guide in which participants need do some
editing. It also explains why some steps are taken and what the expected
result is.