Update Git and GitHub guide
Change around the order to focus on Git first and then to GitHub
when all the Git commands are discussed.

The `rails new` command also initializes a repository already, so the
guides doesn't need a step for this.

There's one problem with this guide still: authentication.
GitHub no longer supports password authentication and this guide
does not address how to authenticate with GitHub right now.
This will addressed later.