Add Mac option in OS switcher for instructions
Add Mac/macOS as its own option in the Operating System switcher.
Sometimes the instructions for Linux and Mac a just a bit different and
it would help to have a separate tab with instructions for Mac.

For pages that have no specific Mac instructions, adding the `mac` class
to the `nix` instructions div-element is enough to make it work for both
Operating Systems.