Automate the page detection for Table of Contents
Instead of relying on a local variable `page` to be given to the
include of the `other-guides` file, detect which page is rendered with
the `page.permalink` variable.

This makes including the other guides partial less likely to break.

Fix a broken "current page" label for the shoulda matchers page in the
process.