Remove the other design guide
It goes into how CSS works and the guides so far have focussed on using
Bootstrap for styling. Adding CSS to it will have unintended results.
Remove it to avoid confusion.

If we do want to address CSS directly (not through Bootstrap), I
recommend adding a more in-depth guide focussing on CSS.