Reorganize the guides
Move some guides to a "more guides" type of section. These are topics
that don't necessarily need to be done in order in the workshop, but are
topics people can look at in any order.

This gives the workshop more focus and a shorter track that can be
completed in a day hopefully.