Upgrade Jekyll, move away from github-pages gem
This website is deployed using GitHub Actions now, and is not limited to
using the github-pages gem to deploy the website to GitHub Pages.

The github-pages gem has its Jekyll dependency locked to version 3.9.3.
Jekyll version 4.3.2 is out, and we can use that directly.

Only the jekyll-redirect-plugin is needed from the github-pages
dependency collection. It's for one redirect, so we can also remove it
when it becomes a problem upgrading Jekyll further.